Remove unsigned long int

+++ 0x05-pointers_arrays_strings/3-puts.c
@@ -8,7 +8,7 @@
  */
 void _puts(char *str)
 {
-	unsigned long int i;
+	 int i;
 
 	for (i = 0; i < 64; i++)
 	{
